Chapter 3: Fix in select_everything, the last example
In the "select everything" section, we renamed `client` to `socket`
since it could also be the server that was returned, in which case we
want to call `accept`.

Anyway, we had to rename the variable in the `puts` call so that the
server does not crash if it receives an empty response (like when the
client sends an empty line with `nc`).

That's it!