Dict: Fix bugs in the `add` method
Separate add from set because there was a bug where calling `add` during
rehashing could lead to a duplicate. If the entry given to add matched
an already existing key in the dictionary that was not yet rehashed, add
would add it to the rehashing table, leading to a duplicate.

This is a problem because we don't check for equality when rehashing,
this would be O(n) since the linked lists in the buckets are not
sorted.

So instead we now have the `set` method, which first calls `get_entry`
to check if the key is present or not, update the entry if it is, add it
otherwise, and the add method which refuses to add a new key if it's
already in either of the tables, main or rehashing.